+++ UpgradeWorld/actions/objects/RefreshObjects.cs
@@ -56,7 +56,7 @@ public class RefreshObjects : EntityOperation
   }
   private void Execute(List<string> ids, DataParameters args)
   {
-    var prefabs = GetPrefabs(ids);
+    var prefabs = GetPrefabs(ids, args.Types);
     var zdos = GetZDOs(args, prefabs);
     var total = 0;
     var counts = prefabs.ToDictionary(prefab => prefab, prefab => 0);