+++ UpgradeWorld/actions/objects/RemoveObjects.cs
@@ -10,7 +10,7 @@ public class RemoveObjects : EntityOperation
   }
   private void Remove(IEnumerable<string> ids, DataParameters args)
   {
-    var prefabs = GetPrefabs(ids);
+    var prefabs = GetPrefabs(ids, args.Types);
     var zdos = GetZDOs(args, prefabs);
     var total = 0;
     var counts = prefabs.ToDictionary(prefab => prefab, prefab => 0);