+++ UpgradeWorld/actions/objects/SearchChests.cs
@@ -23,7 +23,7 @@ public class SearchChests : EntityOperation
   }
   private void Search(IEnumerable<string> ids, DataParameters args)
   {
-    var prefabs = GetPrefabs(ids);
+    var prefabs = GetPrefabs(ids, args.Types);
     var zdos = GetZDOs(args);
 
     var zs = ZNetScene.instance;